[#26] reducer: a 7.13 reducer to host new changes
There are now 3 different reducers, the first one takes care of the
common operation that can be performed on to the state of the UI, for
instance, you'll find in there the elements to configure if the user is
using the yaml view or the form view, what version of the broker the
user want to deploy and if there are some unsaved changes in the page.

Then there are the other 2 reducers, the one for 7.12 that contains all
the logic we had before and the one for 7.13 that will contain
extensions for the 7.12 broker.

In the code you'll have to explicitly use the operation from the reducer
you want to target, ArtemisGlobalReducerOperations for the interface
related one, ArtemisReducerOperations712 for 7.12 and
ArtemisReducerOperations713 for 7.13.