[NO-ISSUE] formview: updating the style
This commit improves the UI by making the center zone scrollable, which
means that if the user has *a lot* of acceptors or connectors, they will
still see the create button at the bottom of the page and the header as
well.

This is an improvement since the main source of feedback for a missing
field is the create button that is disabled.

Also, when the user tries to create a broker with an invalid name (with
CAPS for instance) they will see the error message in the screen instead
of having to scroll back up to have access to the information.