[#21] broker creation: ca don't impact preset
Setting the trust secret should not have any impact on the preset. There
was no warning associated to that field and therefore it used to be a
bug.

Now the user can customize the field without loosing the preset if it
was set. The unit test has been updated accordingly.