[#43] fix loading the resources
The secrets in the resources were not loading due to a strange
interaction between two unrelated calls to the k8sWatchResources hook.
Setting list=False on one made the other also behave as such, even
though its list parameter was set to true.

To address this issue we change flighty the behavior of our hook to
retrieve the broker's secret.

Additionally, this commit fixes the resources container by removing the
use of a useEffect.